Move skills installer under ai command
