+++ cmd/skills_test.go
@@ -12,11 +12,10 @@ import (
 func TestSkillsCommand(t *testing.T) {
 	t.Parallel()
 
-	command := Skills()
+	command := AISkills()
 
 	assert.Equal(t, "skills", command.Name)
-	require.Len(t, command.Commands, 1)
-	assert.Equal(t, "init", command.Commands[0].Name)
+	assert.Empty(t, command.Commands)
 }
 
 func TestLoadBundledBruinSkillsFromMarkdownFiles(t *testing.T) {
@@ -25,13 +24,7 @@ func TestLoadBundledBruinSkillsFromMarkdownFiles(t *testing.T) {
 	skills, err := loadBundledBruinSkills()
 	require.NoError(t, err)
 
-	var skill bundledSkill
-	for _, bundled := range skills {
-		if bundled.Name == "bruin-semantic-layer" {
-			skill = bundled
-			break
-		}
-	}
+	skill := findBundledSkill(t, skills, "bruin-semantic-layer")
 
 	assert.Equal(t, "bruin-semantic-layer", skill.Name)
 
@@ -50,6 +43,9 @@ func TestLoadBundledBruinSkillsFromMarkdownFiles(t *testing.T) {
 	require.True(t, ok)
 	assert.Equal(t, skill.Name, metadata.Name)
 	assert.NotEmpty(t, skill.Files)
+
+	selfHealSkill := findBundledSkill(t, skills, "pipeline-diagnose")
+	assert.Equal(t, []aiConnectionType{aiConnectionBruinCloud, aiConnectionGitHub}, selfHealSkill.RequiredConnections)
 }
 
 func TestSkillsInitCommand_RunCreatesAgentsHomeByDefault(t *testing.T) {
@@ -59,8 +55,8 @@ func TestSkillsInitCommand_RunCreatesAgentsHomeByDefault(t *testing.T) {
 	result, err := (SkillsInitCommand{}).Run(root)
 	require.NoError(t, err)
 
-	require.Len(t, result.Skills, 1)
-	skill := result.Skills[0]
+	require.Len(t, result.Skills, 7)
+	skill := findSkillResult(t, result, "bruin-semantic-layer")
 	assert.Equal(t, skillStatusInstalled, skill.Status)
 	assert.Equal(t, filepath.Join(root, ".agents", "skills"), result.PrimarySkillsDir)
 	assert.Equal(t, filepath.Join(root, ".agents", "skills", "bruin-semantic-layer"), skill.Path)
@@ -72,6 +68,21 @@ func TestSkillsInitCommand_RunCreatesAgentsHomeByDefault(t *testing.T) {
 	assert.True(t, os.IsNotExist(err))
 }
 
+func TestSkillsInitCommand_RunWithOptionsInstallsSelectedSkill(t *testing.T) {
+	t.Parallel()
+
+	root := t.TempDir()
+	result, err := (SkillsInitCommand{}).RunWithOptions(t.Context(), root, SkillsInstallOptions{
+		SkillNames: []string{"bruin-semantic-layer"},
+	})
+	require.NoError(t, err)
+
+	require.Len(t, result.Skills, 1)
+	assert.Equal(t, "bruin-semantic-layer", result.Skills[0].Name)
+	assert.FileExists(t, filepath.Join(root, ".agents", "skills", "bruin-semantic-layer", "SKILL.md"))
+	require.NoDirExists(t, filepath.Join(root, ".agents", "skills", "pipeline-diagnose"))
+}
+
 func TestSkillsInitCommand_RunUsesClaudeWhenOnlyClaudeExists(t *testing.T) {
 	t.Parallel()
 
@@ -81,9 +92,10 @@ func TestSkillsInitCommand_RunUsesClaudeWhenOnlyClaudeExists(t *testing.T) {
 	result, err := (SkillsInitCommand{}).Run(root)
 	require.NoError(t, err)
 
-	require.Len(t, result.Skills, 1)
+	require.Len(t, result.Skills, 7)
 	assert.Equal(t, filepath.Join(root, ".claude", "skills"), result.PrimarySkillsDir)
-	assert.Equal(t, filepath.Join(root, ".claude", "skills", "bruin-semantic-layer"), result.Skills[0].Path)
+	semanticSkill := findSkillResult(t, result, "bruin-semantic-layer")
+	assert.Equal(t, filepath.Join(root, ".claude", "skills", "bruin-semantic-layer"), semanticSkill.Path)
 	assert.FileExists(t, filepath.Join(root, ".claude", "skills", "bruin-semantic-layer", "SKILL.md"))
 
 	_, err = os.Stat(filepath.Join(root, ".agents"))
@@ -100,8 +112,8 @@ func TestSkillsInitCommand_RunLinksClaudeToAgentsWhenBothHomesExist(t *testing.T
 	result, err := (SkillsInitCommand{}).Run(root)
 	require.NoError(t, err)
 
-	require.Len(t, result.Skills, 1)
-	skill := result.Skills[0]
+	require.Len(t, result.Skills, 7)
+	skill := findSkillResult(t, result, "bruin-semantic-layer")
 	require.Len(t, skill.Links, 1)
 
 	primarySkillPath := filepath.Join(root, ".agents", "skills", "bruin-semantic-layer")
@@ -145,8 +157,8 @@ old marker
 	result, err := (SkillsInitCommand{}).Run(root)
 	require.NoError(t, err)
 
-	require.Len(t, result.Skills, 1)
-	assert.Equal(t, skillStatusUpdated, result.Skills[0].Status)
+	semanticSkill := findSkillResult(t, result, "bruin-semantic-layer")
+	assert.Equal(t, skillStatusUpdated, semanticSkill.Status)
 
 	content := readRequiredFile(t, filepath.Join(skillPath, "SKILL.md"))
 	assert.Contains(t, content, "name: bruin-semantic-layer")
@@ -162,13 +174,17 @@ func TestSkillsInitCommand_RunKeepsCurrentSkillContent(t *testing.T) {
 
 	firstResult, err := (SkillsInitCommand{}).Run(root)
 	require.NoError(t, err)
-	require.Len(t, firstResult.Skills, 1)
-	assert.Equal(t, skillStatusInstalled, firstResult.Skills[0].Status)
+	require.Len(t, firstResult.Skills, 7)
+	for _, skill := range firstResult.Skills {
+		assert.Equal(t, skillStatusInstalled, skill.Status, skill.Name)
+	}
 
 	secondResult, err := (SkillsInitCommand{}).Run(root)
 	require.NoError(t, err)
-	require.Len(t, secondResult.Skills, 1)
-	assert.Equal(t, skillStatusCurrent, secondResult.Skills[0].Status)
+	require.Len(t, secondResult.Skills, 7)
+	for _, skill := range secondResult.Skills {
+		assert.Equal(t, skillStatusCurrent, skill.Status, skill.Name)
+	}
 }
 
 func TestSkillsInitCommand_RunUpdatesChangedSkillEvenWithNewerVersion(t *testing.T) {
@@ -190,9 +206,39 @@ newer marker
 	result, err := (SkillsInitCommand{}).Run(root)
 	require.NoError(t, err)
 
-	require.Len(t, result.Skills, 1)
-	assert.Equal(t, skillStatusUpdated, result.Skills[0].Status)
+	semanticSkill := findSkillResult(t, result, "bruin-semantic-layer")
+	assert.Equal(t, skillStatusUpdated, semanticSkill.Status)
 	assert.NotEqual(t, newerSkill, readRequiredFile(t, filepath.Join(skillPath, "SKILL.md")))
+}
+
+func TestSkillsInitCommand_RunWithOptionsSelfHealAddsConnections(t *testing.T) {
+	t.Parallel()
+
+	root := t.TempDir()
+	result, err := (SkillsInitCommand{}).RunWithOptions(t.Context(), root, SkillsInstallOptions{
+		SkillNames: []string{"pipeline-diagnose"},
+		ConnectionResolver: func(connection aiConnectionType) (bool, error) {
+			return true, nil
+		},
+	})
+	require.NoError(t, err)
+
+	require.Len(t, result.Skills, 1)
+	pipelineSkill := findSkillResult(t, result, "pipeline-diagnose")
+	assert.Equal(t, "pipeline-diagnose", pipelineSkill.Name)
+	assert.FileExists(t, filepath.Join(root, ".agents", "skills", "pipeline-diagnose", "SKILL.md"))
+	assert.ElementsMatch(t, []string{"bruin", "github"}, result.ConnectionsAdded)
+
+	cm := readTestBruinConfig(t, filepath.Join(root, ".bruin.yml"))
+	env := cm.Environments["default"]
+	require.Len(t, env.Connections.BruinCloud, 1)
+	assert.Equal(t, "bruin-cloud", env.Connections.BruinCloud[0].Name)
+	assert.Equal(t, "${BRUIN_CLOUD_API_TOKEN}", env.Connections.BruinCloud[0].APIToken)
+	require.Len(t, env.Connections.GitHub, 1)
+	assert.Equal(t, "github", env.Connections.GitHub[0].Name)
+	assert.Equal(t, "${GITHUB_TOKEN}", env.Connections.GitHub[0].AccessToken)
+	assert.Equal(t, "<github-owner>", env.Connections.GitHub[0].Owner)
+	assert.Equal(t, "<github-repo>", env.Connections.GitHub[0].Repo)
 }
 
 func readRequiredFile(t *testing.T, path string) string {
@@ -201,4 +247,28 @@ func readRequiredFile(t *testing.T, path string) string {
 	content, err := os.ReadFile(path)
 	require.NoError(t, err)
 	return string(content)
+}
+
+func findBundledSkill(t *testing.T, skills []bundledSkill, name string) bundledSkill {
+	t.Helper()
+
+	for _, skill := range skills {
+		if skill.Name == name {
+			return skill
+		}
+	}
+	t.Fatalf("bundled skill %s not found", name)
+	return bundledSkill{}
+}
+
+func findSkillResult(t *testing.T, result *SkillsInitResult, name string) SkillInstallResult {
+	t.Helper()
+
+	for _, skill := range result.Skills {
+		if skill.Name == name {
+			return skill
+		}
+	}
+	t.Fatalf("skill result %s not found", name)
+	return SkillInstallResult{}
 }
