+++ main.go
@@ -69,7 +69,6 @@ func main() {
 			cmd.CleanCmd(),
 			cmd.Format(&isDebug),
 			cmd.AI(&isDebug),
-			cmd.Skills(),
 			cmd.Docs(),
 			cmd.Init(),
 			cmd.Internal(),
