+++ templates/templates.go
@@ -6,6 +6,4 @@ import (
 
 //go:embed *
 //go:embed */.bruin.yml
-//go:embed ai/agents-md/AGENTS.md
-//go:embed ai/skill-self-heal/.agents/skills/*/SKILL.md
 var Templates embed.FS
